historical-data-migrator-lambda/requirements.txt
boto3
coralogix_logger
